Make #exif and #iptc 'magic' attributes.
Created a Dragonfly Analyser to handle extraction of IPTC and EXIF image data. Extracted data is auto-populated into the #image_exif and #image_iptc attributes. The rest of the Photo model was updated to account for this.

Also refactored handling population of #release_date... see comments in app/models/photo.rb for details.