Enables lazy instantiation of Permalink.
Delegating :slug to :permalink caused a problem, since attributes are mass-assigned to models _before_ the after_initialization callback is invoked. So when admin forms posted the :slug attribute for a model, the delegation ran into a nil Permalink.

This corrects that condition by building Permalink on-demand if one does not already exist.